[working] ability to select which of the show/edit/delete buttons should show on a data table

+++ app/Base/Helpers/AdminHelper.php
@@ -9,8 +9,11 @@ if (!function_exists('get_ops')) {
      * @param $class
      * @return string
      */
-    function get_ops($resource, $id, $class = "btn", array $ops)
+    function get_ops($resource, $id, $class = "btn", array $showable_ops)
     {
+        // No ops? No list!
+        if (!count($showable_ops)) return "";
+
         if ($class=="btn") {
             $show_class = "btn btn-xs bg-navy";
             $edit_class = "btn btn-xs bg-olive";
@@ -23,25 +26,36 @@ if (!function_exists('get_ops')) {
         $show_path = route('admin.'.$resource.'.show', ['id' => $id]);
         $edit_path = route('admin.'.$resource.'.edit', ['id' => $id]);
         $delete_path = route('admin.'.$resource.'.destroy', ['id' => $id]);
+
         $ops  = '<ul class="list-inline no-margin-bottom">';
-        $ops .=  '<li>';
-        $ops .=  '<a class="'.$show_class.'" href="'.$show_path.'">
-                  <i class="fa fa-search"></i>
-                  '.trans('admin.ops.show').'</a>';
-        $ops .=  '</li>';
-        $ops .=  '<li>';
-        $ops .=  '<a class="'.$edit_class.'" href="'.$edit_path.'">
-                 <i class="fa fa-pencil-square-o"></i>
-                  '.trans('admin.ops.edit').'</a>';
-        $ops .=  '</li>';
-        $ops .=  '<li>';
-        $ops .= Form::open(['method' => 'DELETE', 'url' => $delete_path]);
-        $ops .= Form::submit('&#xf1f8; ' .trans('admin.ops.delete'), [
-                'onclick' => "return confirm('".trans('admin.ops.confirmation')."');",
-                'class' => $delete_class
-            ]);
-        $ops .= Form::close();
-        $ops .=  '</li>';
+        if (in_array("show", $showable_ops))
+        {
+            $ops .=  '<li>';
+            $ops .=  '<a class="'.$show_class.'" href="'.$show_path.'">
+                      <i class="fa fa-search"></i>
+                      '.trans('admin.ops.show').'</a>';
+            $ops .=  '</li>';
+        }
+
+        if (in_array("edit", $showable_ops)) {
+            $ops .=  '<li>';
+            $ops .=  '<a class="'.$edit_class.'" href="'.$edit_path.'">
+                     <i class="fa fa-pencil-square-o"></i>
+                      '.trans('admin.ops.edit').'</a>';
+            $ops .=  '</li>';
+        }
+
+        if (in_array("delete", $showable_ops))
+        {
+            $ops .=  '<li>';
+            $ops .= Form::open(['method' => 'DELETE', 'url' => $delete_path]);
+            $ops .= Form::submit('&#xf1f8; ' .trans('admin.ops.delete'), [
+                    'onclick' => "return confirm('".trans('admin.ops.confirmation')."');",
+                    'class' => $delete_class
+                ]);
+            $ops .= Form::close();
+            $ops .=  '</li>';
+        }
         $ops .=  '</ul>';
         return $ops;
     }
